// utils/graphql/storefront.graphql.gen.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Any: { input: any; output: any };
  CEP: { input: any; output: any };
  CountryCode: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  EmailAddress: { input: any; output: any };
  Long: { input: any; output: any };
  Upload: { input: any; output: any };
  Uuid: { input: any; output: any };
};
// ...
export type BuyBox = {
  /** List of the possibles installment plans. */
  installmentPlans?: Maybe<Array<Maybe<InstallmentPlan>>>;
  /** Maximum price among sellers. */
  maximumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Minimum price among sellers. */
  minimumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Quantity of offers. */
  quantityOffers?: Maybe<Scalars["Int"]["output"]>;
  /** List of sellers. */
  sellers?: Maybe<Array<Maybe<Seller>>>;
};
// ...
export type HotsiteSubtype =
  /** Hotsite created from a brand. */
  | "BRAND"
  /** Hotsite created from a buy list (lista de compra). */
  | "BUY_LIST"
  /** Hotsite created from a category. */
  | "CATEGORY"
  /** Hotsite created from a portfolio. */
  | "PORTFOLIO";
// ...
export type PartnerSortKeys =
  /** The partner unique identifier. */
  | "ID"
  /** The partner name. */
  | "NAME";

export type PartnerSubtype =
  /** Partner 'client' subtype. */
  | "CLIENT"
  /** Partner 'closed' subtype. */
  | "CLOSED"
  /** Partner 'open' subtype. */
  | "OPEN";
// ...
export type ProductFilterInput = {
  /** The attribute name. */
  field: Scalars["String"]["input"];
  /** The set of values which the result filter item value must be included in. */
  values: Array<InputMaybe<Scalars["String"]["input"]>>;
};
// ...
export type ProductSortKeys =
  /** The applied discount to the product variant price. */
  | "DISCOUNT"
  /** The product name. */
  | "NAME"
  /** The product variant price. */
  | "PRICE"
  /** Sort in a random way. */
  | "RANDOM"
  /** The date the product was released. */
  | "RELEASE_DATE"
  /** The sales number on a period of time. */
  | "SALES"
  /** The quantity in stock of the product variant. */
  | "STOCK";
// ...
export type SortDirection =
  /** The results will be sorted in an ascending order. */
  | "ASC"
  /** The results will be sorted in an descending order. */
  | "DESC";
// ...
export type UriKind =
  | "BUY_LIST"
  | "HOTSITE"
  | "NOT_FOUND"
  | "PARTNER"
  | "PRODUCT"
  | "REDIRECT";
// ...
export type ProductFragment = {
  mainVariant?: boolean | null;
  productName?: string | null;
  productId?: any | null;
  alias?: string | null;
  available?: boolean | null;
  averageRating?: number | null;
  condition?: string | null;
  createdAt?: any | null;
  ean?: string | null;
  id?: string | null;
  minimumOrderQuantity?: number | null;
  productVariantId?: any | null;
  parentId?: any | null;
  sku?: string | null;
  numberOfVotes?: number | null;
  stock?: any | null;
  variantName?: string | null;
  variantStock?: any | null;
  collection?: string | null;
  urlVideo?: string | null;
  attributes?: Array<{
    value?: string | null;
    name?: string | null;
    attributeId?: string | null;
    displayType?: string | null;
    id?: string | null;
  } | null> | null;
  productCategories?: Array<{
    id: number;
    name?: string | null;
    url?: string | null;
    hierarchy?: string | null;
    main: boolean;
    googleCategories?: string | null;
  } | null> | null;
  informations?: Array<{
    title?: string | null;
    value?: string | null;
    type?: string | null;
  } | null> | null;
  images?: Array<{
    url?: string | null;
    fileName?: string | null;
    print: boolean;
  } | null> | null;
  prices?: {
    discountPercentage: any;
    discounted: boolean;
    listPrice?: any | null;
    multiplicationFactor: number;
    price: any;
    bestInstallment?: {
      discount: boolean;
      displayName?: string | null;
      fees: boolean;
      name?: string | null;
      number: number;
      value: any;
    } | null;
    installmentPlans?: Array<{
      displayName?: string | null;
      name?: string | null;
      installments?: Array<{
        discount: boolean;
        fees: boolean;
        number: number;
        value: any;
      } | null> | null;
    } | null> | null;
    priceTables?: Array<{
      discountPercentage: any;
      id: any;
      listPrice?: any | null;
      price: any;
    } | null> | null;
    wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
  } | null;
  productBrand?: {
    fullUrlLogo?: string | null;
    logoUrl?: string | null;
    name?: string | null;
    alias?: string | null;
  } | null;
  seller?: { name?: string | null } | null;
  similarProducts?: Array<{
    alias?: string | null;
    image?: string | null;
    imageUrl?: string | null;
    name?: string | null;
  } | null> | null;
  promotions?: Array<{
    content?: string | null;
    disclosureType?: string | null;
    id: any;
    fullStampUrl?: string | null;
    stamp?: string | null;
    title?: string | null;
  } | null> | null;
  buyBox?: BuyBox | null;
};
// ...
export type GetUrlQueryVariables = Exact<{
  url: Scalars["String"]["input"];
}>;

export type GetUrlQuery = {
  uri?: {
    hotsiteSubtype?: HotsiteSubtype | null;
    kind: UriKind;
    partnerSubtype?: PartnerSubtype | null;
    productAlias?: string | null;
    productCategoriesIds?: Array<number> | null;
    redirectCode?: string | null;
    redirectUrl?: string | null;
  } | null;
};
// ...
export type HotsiteQueryVariables = Exact<{
  url?: InputMaybe<Scalars["String"]["input"]>;
  filters?: InputMaybe<
    Array<InputMaybe<ProductFilterInput>> | InputMaybe<ProductFilterInput>
  >;
  limit?: InputMaybe<Scalars["Int"]["input"]>;
  maximumPrice?: InputMaybe<Scalars["Decimal"]["input"]>;
  minimumPrice?: InputMaybe<Scalars["Decimal"]["input"]>;
  onlyMainVariant?: InputMaybe<Scalars["Boolean"]["input"]>;
  offset?: InputMaybe<Scalars["Int"]["input"]>;
  sortDirection?: InputMaybe<SortDirection>;
  sortKey?: InputMaybe<ProductSortKeys>;
  partnerAccessToken?: InputMaybe<Scalars["String"]["input"]>;
  ignoreDisplayRules?: boolean;
  hotsiteId?: InputMaybe<Scalars["Int"]["input"]>;
}>;

export type HotsiteQuery = {
  result?: {
    endDate?: any | null;
    expression?: string | null;
    id?: string | null;
    name?: string | null;
    pageSize: number;
    startDate?: any | null;
    subtype?: HotsiteSubtype | null;
    template?: string | null;
    url?: string | null;
    hotsiteId: any;
    aggregations?: {
      maximumPrice: any;
      minimumPrice: any;
      filters?: Array<{
        field?: string | null;
        origin?: string | null;
        values?: Array<{
          name?: string | null;
          quantity: number;
        } | null> | null;
      } | null> | null;
      priceRanges?: Array<{
        quantity: number;
        range?: string | null;
      } | null> | null;
    } | null;
    productsByOffset?: {
      page: number;
      pageSize: number;
      totalCount: number;
      items?: Array<{
        mainVariant?: boolean | null;
        productName?: string | null;
        productId?: any | null;
        alias?: string | null;
        available?: boolean | null;
        averageRating?: number | null;
        condition?: string | null;
        createdAt?: any | null;
        ean?: string | null;
        id?: string | null;
        minimumOrderQuantity?: number | null;
        productVariantId?: any | null;
        parentId?: any | null;
        sku?: string | null;
        numberOfVotes?: number | null;
        stock?: any | null;
        variantName?: string | null;
        variantStock?: any | null;
        collection?: string | null;
        urlVideo?: string | null;
        attributes?: Array<{
          value?: string | null;
          name?: string | null;
        } | null> | null;
        productCategories?: Array<{
          id: number;
          name?: string | null;
          url?: string | null;
          hierarchy?: string | null;
          main: boolean;
          googleCategories?: string | null;
        } | null> | null;
        informations?: Array<{
          title?: string | null;
          value?: string | null;
          type?: string | null;
        } | null> | null;
        images?: Array<{
          url?: string | null;
          fileName?: string | null;
          print: boolean;
        } | null> | null;
        prices?: {
          discountPercentage: any;
          discounted: boolean;
          listPrice?: any | null;
          multiplicationFactor: number;
          price: any;
          bestInstallment?: {
            discount: boolean;
            displayName?: string | null;
            fees: boolean;
            name?: string | null;
            number: number;
            value: any;
          } | null;
          installmentPlans?: Array<{
            displayName?: string | null;
            name?: string | null;
            installments?: Array<{
              discount: boolean;
              fees: boolean;
              number: number;
              value: any;
            } | null> | null;
          } | null> | null;
          priceTables?: Array<{
            discountPercentage: any;
            id: any;
            listPrice?: any | null;
            price: any;
          } | null> | null;
          wholesalePrices?: Array<{
            price: any;
            quantity: number;
          } | null> | null;
        } | null;
        productBrand?: {
          fullUrlLogo?: string | null;
          logoUrl?: string | null;
          name?: string | null;
          alias?: string | null;
        } | null;
        seller?: { name?: string | null } | null;
        similarProducts?: Array<{
          alias?: string | null;
          image?: string | null;
          imageUrl?: string | null;
          name?: string | null;
        } | null> | null;
        promotions?: Array<{
          content?: string | null;
          disclosureType?: string | null;
          id: any;
          fullStampUrl?: string | null;
          stamp?: string | null;
          title?: string | null;
        } | null> | null;
        buyBox?: BuyBox | null;
      } | null> | null;
    } | null;
    breadcrumbs?: Array<{
      link?: string | null;
      text?: string | null;
    } | null> | null;
    seo?: Array<{
      content?: string | null;
      httpEquiv?: string | null;
      name?: string | null;
      scheme?: string | null;
      type?: string | null;
    } | null> | null;
    sorting?: {
      direction?: SortDirection | null;
      field?: ProductSortKeys | null;
    } | null;
  } | null;
};
// ...
export type GetPartnersQueryVariables = Exact<{
  first?: InputMaybe<Scalars["Int"]["input"]>;
  last?: InputMaybe<Scalars["Int"]["input"]>;
  names?: InputMaybe<
    | Array<InputMaybe<Scalars["String"]["input"]>>
    | InputMaybe<Scalars["String"]["input"]>
  >;
  priceTableIds?: InputMaybe<
    Array<Scalars["Int"]["input"]> | Scalars["Int"]["input"]
  >;
  sortDirection?: SortDirection;
  sortKey?: PartnerSortKeys;
  before?: InputMaybe<Scalars["String"]["input"]>;
  alias?: InputMaybe<
    | Array<InputMaybe<Scalars["String"]["input"]>>
    | InputMaybe<Scalars["String"]["input"]>
  >;
  after?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type GetPartnersQuery = {
  partners?: {
    edges?: Array<{
      node?: {
        partnerId: any;
        priceTableId: number;
        portfolioId: number;
        type?: string | null;
        startDate: any;
        endDate: any;
        name?: string | null;
        alias?: string | null;
        fullUrlLogo?: string | null;
        origin?: string | null;
        partnerAccessToken?: string | null;
      } | null;
    }> | null;
  } | null;
};